:recycle: move cached_slot_property to private
