+++ discord/sticker.py
@@ -32,7 +32,8 @@ from .asset import Asset, AssetMixin
 from .enums import StickerFormatType, StickerType, try_enum
 from .errors import InvalidData
 from .mixins import Hashable
-from .utils import MISSING, Undefined, cached_slot_property, find, snowflake_time
+from .utils import MISSING, Undefined, find, snowflake_time
+from .utils.private import cached_slot_property
 
 __all__ = (
     "StickerPack",
